Rearchitected the code to construct the graph in a separate in file and then graph_server calls a function to get the graph.
